#8 Divide process into launch process and cron job child process
